Updates `parseDomFragment()` to return an `HTMLTemplateElement` and SSR the editable tweet.
This allows a single fetched element to be cloned as necessary and reused in multiple locations. For this example, we SSR the editable tweet component on app load and clone it each time an edit is requested. I'm not convinced this is better than CSR-ing the editable tweet, but it is more interesting which is why I decided to go with the approach.