Adds streamable HTML fragments.
There's a lot of nuance to the implementation here, so I'll mostly let inline comments do the talking. Check out the [new blog post](https://blog.dwac.dev/posts/streamable-html-fragments/) for more info on all the nuances I had to solve along the way and more insight into why it's written the way it is.