Add Firefox support to streaming use case.
Firefox didn't work for streaming use cases and would never trigger the `MutationObserver`. This appears to be because `doc.write()` implicitly calls `doc.open()` if it wasn't explicitly called previously. This implicit `doc.open()` resets the document, detaching `doc.body` and setting it to `null`. When `doc.write()` actually parses the input HTML, it creates a new `doc.body` element and appends nodes to it. However the `MutationObserver` was observing the previous `doc.body`, not the new one, so it does not receive these events.

The solution here is to manually call `doc.open()` and then observe that `doc.body`, so the reset happens before the `MutationObserve` starts. Unfortunately `doc.open()` resets `doc.body` to `null`, so we actually have to manually parse enough HTML to create a new `<body>` tag. This isn't ideal, but good enough for HTML fragments, since `<html>`, `<head>`, and `<body>` tag shouldn't be in the content anyways.

See explanation in https://bugzilla.mozilla.org/show_bug.cgi?id=1811782#c3.